Running games on REPL !
